+++ lib/text_word_highlighter.dart
@@ -1,37 +1,59 @@
 library text_word_highlighter;
 
 import 'package:flutter/material.dart';
+import 'package:text_word_highlighter/utils/highlight_utils.dart';
 import 'package:text_word_highlighter/utils/word_highlight.dart';
 
+///Highlight multiple words inside the text
 class TextWordHighlighter extends StatelessWidget {
+  ///Highlight multiple words inside the text
   TextWordHighlighter(
       {Key? key,
       required this.text,
       required this.wordHighlightList,
       required this.textStyle,
-      this.overflow = TextOverflow.clip})
+      this.overflow = TextOverflow.clip,
+      this.textDirection,
+      this.maxLines,
+      this.locale})
       : super(key: key);
 
-  // The input sentence or string
+  /// The input sentence or string
   final String text;
 
-  // List of WordHighlight that contains index, text style of the word present at the index
+  /// List of WordHighlight that contains index, text style of the word present at the index
   final List<WordHighlight> wordHighlightList;
 
-  // Split the text(sentence) into words and store it into _wordList
+  /// Split the text(sentence) into words and store it into _wordList
   final List<String> _wordList = [];
 
-  // Common text style for all the words in the sentence
+  /// Common text style for all the words in the sentence
   final TextStyle textStyle;
 
-  // text overflow
+  /// text overflow
   final TextOverflow overflow;
+
+  /// Rich text properties
+  final TextAlign textAlign = TextAlign.start;
+  final TextDirection? textDirection;
+  final bool softWrap = true;
+  final double textScaleFactor = 1.0;
+  final TextScaler textScaler = TextScaler.noScaling;
+  final int? maxLines;
+  final Locale? locale;
 
   @override
   Widget build(BuildContext context) {
     // split the sentence into list of words separated by space
     _wordList.addAll(text.split(" "));
     return RichText(
+      textAlign: textAlign,
+      textDirection: textDirection,
+      softWrap: softWrap,
+      textScaleFactor: textScaleFactor,
+      textScaler: textScaler,
+      maxLines: maxLines,
+      locale: locale,
       text: TextSpan(style: textStyle, children: _getTextSpan()),
       overflow: overflow,
     );
@@ -48,8 +70,10 @@ class TextWordHighlighter extends StatelessWidget {
       }
 
       // Get highlighter for the given index
-      final highlighters =
-          wordHighlightList.where((element) => element.wordIndex == i).toList();
+      final highlighters = wordHighlightList
+          .where((wordHighlight) => isHighlightedWord(_wordList,
+              index: i, wordHighlight: wordHighlight))
+          .toList();
 
       // Check if any highlighter exists for the given index
       spanList.add(TextSpan(