+++ CHANGELOG.md
@@ -12,3 +12,20 @@
 * index inside Highlighter renamed to wordIndex, and style as wordStyle.
 * Some know issues(Font size issues) fixed.
 * Space issue fixed.
+
+## 1.0.0
+* Documentation improved
+* Example added
+* Some breaking changes has been done
+* Highlight by index feature added.
+
+## 1.0.1
+* Added Few attributes to TextWordHighlighter as below
+  TextAlign textAlign = TextAlign.start;
+  TextDirection? textDirection;
+  bool softWrap = true;
+  double textScaleFactor = 1.0;
+  TextScaler textScaler = TextScaler.noScaling;
+  int? maxLines;
+  Locale? locale;
+